requirements.txt
flask
authlib
python-dotenv
boto3
pyjwt
